Move serde to a module
